move effect that removes raster symbology on active project change outside homemap
